pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecological-agents"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "psycopg2-binary>=2.9.9",
  "pgvector>=0.3.4",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "aiohttp>=3.9",
  "pydantic>=2.9.0",
  "numpy>=2.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
